backend/requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.1.0
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
